Fixed plugin issue, improved plugin registration
The source code plugin was assuming that the ethersplay directory was symlinked in the plugins directory, but this is annoying to do on Windows, which meant that the plugin didn't work correctly. It's been moved around so that it can import what it needs.

Additionally, I added an is_valid method so that the EVM plugins only display for EVM architectures.