Linting and tracking bbs with Function.session_data
In order to not run across all basic blocks every time the `BinaryDataNotification` callback occurs, the new blocks to explore, as well as the `StackValueAnalysis` object are cached in the session_data dict of the Function.

Additionally, I cleaned up some of the code using flake8. Still some cleanup left to do.